Format project

+++ technician-scheduling-optaplanner-example/src/main/java/org/optaplanner/examples/common/swingui/latitudelongitude/LatitudeLongitudeTranslator.java
@@ -80,11 +80,11 @@ public class LatitudeLongitudeTranslator {
     }
 
     public double translateXToLongitude(int x) {
-        return minimumLongitude + ((((double) x) - innerWidthMargin) * longitudeLength / innerWidth);
+        return minimumLongitude + (((x) - innerWidthMargin) * longitudeLength / innerWidth);
     }
 
     public double translateYToLatitude(double y) {
-        return maximumLatitude - ((((double) y) - innerHeightMargin) * latitudeLength / innerHeight);
+        return maximumLatitude - (((y) - innerHeightMargin) * latitudeLength / innerHeight);
     }
 
     public int getImageWidth() {
